+++ src/base64.ml
@@ -21,12 +21,16 @@ let count_len str =
   done;
   !i + 1
 
+exception Error of string
+
+let[@inline] error descr = raise (Error descr)
+
 let decode input =
   let table = Lazy.force decode_table in
   let real_len = count_len input in
   let padding = (4 - real_len land 3) land 3 in
   if padding = 3 then
-    failwith "Invalid base64: a padding of 3 sextets is invalid";
+    error "Invalid base64: a padding of 3 sextets is invalid";
   (* padding can be either 0, 1, or 2 *)
   let out_len = (real_len + padding) / 4 * 3 - padding in
   let out = Bytes.create out_len in
@@ -37,7 +41,7 @@ let decode input =
     let ch2 = input.%[!i + 1] in
     let sextet1 = table.%(Char.code ch1) in
     let sextet2 = table.%(Char.code ch2) in
-    if sextet1 = 0xff || sextet2 = 0xff then failwith "Invalid base64 character";
+    if sextet1 = 0xff || sextet2 = 0xff then error "Invalid base64 character";
     (* 6 + 2 bits *)
     let octet1 = sextet1 lsl 2
              lor sextet2 lsr 4 in
@@ -46,7 +50,7 @@ let decode input =
     if !i + 2 < real_len then begin
       let ch3 = input.%[!i + 2] in
       let sextet3 = table.%(Char.code ch3) in
-      if sextet3 = 0xff then failwith "Invalid base64 character";
+      if sextet3 = 0xff then error "Invalid base64 character";
       (* 4 + 4 bits *)
       let octet2 = (sextet2 land 0b1111) lsl 4
                lor sextet3 lsr 2 in
@@ -55,7 +59,7 @@ let decode input =
       if !i + 3 < real_len then begin
         let ch4 = input.%[!i + 3] in
         let sextet4 = table.%(Char.code ch4) in
-        if sextet4 = 0xff then failwith "Invalid base64 character";
+        if sextet4 = 0xff then error "Invalid base64 character";
         (* 2 + 6 bits *)
         let octet3 = (sextet3 land 0b11) lsl 6
                  lor sextet4 in