+++ src/base64.mli
@@ -1,4 +1,7 @@
+exception Error of string
+(** [Error description]: Base64 decoding error. *)
+
 val decode : string -> bytes
 (** Decode a base64-encoded string. This can decode base64 both with and
     without padding. The string can end with arbitrary number of [=] characters.
-    @raise Failure in case of invalid base64 *)
+    @raise Error in case of invalid base64 *)