+++ CHANGES.md
@@ -8,6 +8,7 @@
 - - Added whitespace escape sequence: all whitespace (including multiple
     newlines) is removed after `\`
 - - Identifiers cannot start with `r#` anymore
+- `interpret` now raises `Invalid_annotation` instead of `Failure`.
 - Dropped support for OCaml < 4.10.0
 
 ## 0.1.0 (2022-10-01)